+++ CHANGELOG.md
@@ -17,6 +17,7 @@ HumHub Changelog
 - Enh #8105: Fix adding of new individual user and space topics
 - Fix #8107: Hide option "Update Space memberships" on administrative group edit form
 - Enh #8106: ensure migrations are run after reinstallation of enabled modules
+- Fix #8118: Keep list start number on RichText rendering
 
 1.18.2 (March 22, 2026)
 -----------------------